requirements.txt
filelock==3.16.1
fsspec==2024.12.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
